+++ src/test/java/com/wikia/webdriver/elements/oasis/pages/TemplateEditPage.java
@@ -1,4 +1,4 @@
-package com.wikia.webdriver.pageobjectsfactory.pageobject.template;
+package com.wikia.webdriver.elements.oasis.pages;
 
 import org.openqa.selenium.WebDriver;
